refactor: Consolidate workflow scheduling by removing WorkflowSchedulerService and updating imports
